requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
websockets
httpx
requests